fix: make send queue safe
In which I finally fix the subtle soundness issue that `.start_send()`
must not be called otherwise you'll indefinitely grow the queue in the
channel reguardless of the buffer size.

The old send method is now called `force_send`, it is at this point
still used in restart and stop usecases.